% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{biologer.login}
\alias{biologer.login}
\title{Sets the Biologer server and access token}
\usage{
biologer.login(server = "rs", token)
}
\arguments{
\item{server}{Alias of the Biologer server. Available values are: 'rs', 'hr', 'bs', and 'me'.}

\item{token}{String representing the access token obtained from the user preferences
on the Biologer web page.}
}
\value{
Invisible logical calue (TRUE) if the server and token are successfully saved.
The function stops the execution and displays the error if the server/token is invalid.
}
\description{
This function allows user to set the Biologer server (sr, hr, ba or me) and
access token for the API calls. Obtain access token from User > PreferencesAPI Tokens. Your server and access token will be saved in local '~/.Renviron'file and loaded automatically in your R session.
}
\note{
To obtain access token, the user must be logged in to Biologer web page using your browser.
Here you can obtain the token by selecting User > Preferences > API Tokens > Generate token.
The variables will be set automatically for current session, but you need to restart R session
for the changes to became permanent.
}
\examples{
\dontrun{
# Example: Setting the access token for Croatian Biologer server
biologer.login(
  server = "hr"
  token = "TOKEN_OBRAINED FROM THE BROWSER",
)

# Taken and URL are not availavle in internal package functions.
}

}
