% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{collapse_photo_urls}
\alias{collapse_photo_urls}
\title{Collapse nested 'photos' data frame into a semicolon-separated string of URLs.}
\usage{
collapse_photo_urls(photos_df)
}
\arguments{
\item{photos_df}{A data frame containing 'url' column, or an empty list/data frame.}
}
\description{
Collapse nested 'photos' data frame into a semicolon-separated string of URLs.
}
\keyword{internal}
