% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{collapse_stages}
\alias{collapse_stages}
\title{Collapse nested 'stages' data frame into a semicolon-separated string of stage names.}
\usage{
collapse_stages(stages_df)
}
\arguments{
\item{stages_df}{A data frame containing 'name' column, or an empty list/data frame.}
}
\description{
Collapse nested 'stages' data frame into a semicolon-separated string of stage names.
}
\keyword{internal}
