% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{collapse_translations}
\alias{collapse_translations}
\title{Collapse nested 'translations' data frame into a semicolon-separated string of locale=name pairs.}
\usage{
collapse_translations(translations_df)
}
\arguments{
\item{translations_df}{A data frame containing 'locale' and 'native_name' columns, or an empty list/data frame.}
}
\description{
Collapse nested 'translations' data frame into a semicolon-separated string of locale=name pairs.
}
\keyword{internal}
