% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{collapse_types}
\alias{collapse_types}
\title{Collapse nested 'types' data frame into a semicolon-separated string of names.}
\usage{
collapse_types(type_df)
}
\arguments{
\item{type_df}{A data frame containing 'name' column, or an empty list/data frame.}
}
\description{
Collapse nested 'types' data frame into a semicolon-separated string of names.
}
\keyword{internal}
