% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{display_progress}
\alias{display_progress}
\title{Display dynamic, single-line progress update in the console.}
\usage{
display_progress(records_fetched, total_records)
}
\arguments{
\item{records_fetched}{The total number of records fetched so far.}

\item{total_records}{The total number of records expected (or NULL if unknown).}

\item{current_count}{The number of records fetched in the current page (optional, for debugging).}
}
\description{
Display dynamic, single-line progress update in the console.
}
\keyword{internal}
