% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{extract_taxa_from_activity}
\alias{extract_taxa_from_activity}
\title{Extract the old taxon name from a single 'properties' list element.}
\usage{
extract_taxa_from_activity(property_list)
}
\arguments{
\item{property_list}{A single list element from the 'properties' column.}
}
\value{
The old taxon name (character string) or an empty string if not found.
}
\description{
It searches for 'properties$old$taxon$label'.
}
\keyword{internal}
