% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{format_authors}
\alias{format_authors}
\title{Formats a list of publication authors into a standard citation string.}
\usage{
format_authors(authors_list)
}
\arguments{
\item{authors_list}{A list column element containing a data frame of authors
with 'last_name' and 'first_name' columns.}
}
\description{
Formats a list of publication authors into a standard citation string.
}
\keyword{internal}
