% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{get_data_from_api}
\alias{get_data_from_api}
\title{Fetch Paginated Observations and Save/Resume using Timestamp}
\usage{
get_data_from_api(
  api_url,
  token,
  filename = NULL,
  columns_to_keep = NULL,
  verbose = TRUE
)
}
\arguments{
\item{api_url}{URL of the Biologer API used to access the data.}

\item{token}{JWT authorization token.}

\item{filename}{Character string for the path to save the final data
(e.g., "observations.csv").}

\item{columns_to_keep}{Select columns to keep in the final dataset. Server
returns many columns, which are not required for RBiologer to work.}
}
\description{
Resumes the download from the last successfully fetched record's timestamp.
}
\keyword{internal}
