% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{get_data_from_simple_api}
\alias{get_data_from_simple_api}
\title{Fetch simple data from the API that does not require pagination and timestamps}
\usage{
get_data_from_simple_api(api_url, token, filename = NULL)
}
\arguments{
\item{api_url}{URL of the Biologer API used to access the data.}

\item{token}{JWT authorization token.}
}
\description{
Fetch simple data from the API that does not require pagination and timestamps
}
\keyword{internal}
