% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{get_literature_observations}
\alias{get_literature_observations}
\title{Fetch Paginated Literature Observations and Save/Resume using Timestamp}
\usage{
get_literature_observations(base_url, token, filename = NULL)
}
\arguments{
\item{base_url}{Base URL of the API.}

\item{token}{JWT authorization token.}

\item{filename}{Character string for the path to save the final data
(e.g., "literature_observations.csv").}
}
\description{
Resumes the download from the last successfully fetched record's timestamp.
}
\keyword{internal}
