% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{get_multiple_env_vars}
\alias{get_multiple_env_vars}
\title{Internal helper function to retrieve multiple environment variables based on a prefix}
\usage{
get_multiple_env_vars(prefix, server_keys)
}
\description{
Internal helper function to retrieve multiple environment variables based on a prefix
}
\keyword{internal}
