% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{get_storage_path}
\alias{get_storage_path}
\title{Get the internal storage directory for the package}
\usage{
get_storage_path(package = "RBiologer")
}
\description{
Get the internal storage directory for the package
}
\keyword{internal}
