% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{get_taxon_by_id}
\alias{get_taxon_by_id}
\title{Get Taxon Data by ID from Biologer API}
\usage{
get_taxon_by_id(taxon_id, server = NULL, verbose = TRUE)
}
\arguments{
\item{taxon_id}{Numeric or Character. The ID of the taxon you want to fetch.}

\item{server}{Character. The name of the server (e.g., "rs", "hr", "ba", "me").
If NULL, it will try to use the first available token.}

\item{verbose}{Logical. Print progress and status messages?}
}
\value{
A list containing the taxon data, or an error if the request fails.
}
\description{
Get Taxon Data by ID from Biologer API
}
