% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{is_taxon_change_event}
\alias{is_taxon_change_event}
\title{Check if a single activity record is a taxon change event.}
\usage{
is_taxon_change_event(property_list)
}
\arguments{
\item{property_list}{A single list element from the 'properties' column.}
}
\value{
TRUE if it's a taxon change, FALSE otherwise.
}
\description{
Check if a single activity record is a taxon change event.
}
\keyword{internal}
