% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_biologer.R
\name{plot.hr}
\alias{plot.hr}
\title{Plot RH_2018 Shapefile}
\usage{
plot.hr()
}
\value{
A ggplot object showing the RH_2018 map.
}
\description{
Loads and plots the RH_2018 shapefile included in the package.
}
