% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_biologer.R
\name{plot.zupanija}
\alias{plot.zupanija}
\title{Plot Zupanije JSON}
\usage{
plot.zupanija()
}
\value{
A ggplot object showing the Zupanije map.
}
\description{
Loads and plots the Zupanije.json file included in the package.
}
