% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{remove_from_renviron}
\alias{remove_from_renviron}
\title{Internal function for removing variables from .Renviron file}
\usage{
remove_from_renviron(key)
}
\description{
Internal function for removing variables from .Renviron file
}
\keyword{internal}
