% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biologer.R
\name{set_per_page}
\alias{set_per_page}
\title{Set or Update the 'per_page' Query Parameter in a URL}
\usage{
set_per_page(url, per.page = 1000)
}
\arguments{
\item{url}{A character string representing the URL to be modified.
Can be the initial base API path or a 'next' link from a previous API response.}

\item{per.page}{An integer specifying the desired number of records per page.
Defaults to 1000.}
}
\value{
A character string representing the modified URL with the 'per_page' parameter set.
Returns \code{NULL} if the input \code{url} is \code{NULL}.
}
\description{
Ensures that a given URL for the Biologer API includes the 'per_page'
query parameter, setting it to a specified value (defaulting to 1000).
}
\details{
This function safely sets the pagination size for API requests by:
\enumerate{
\item Appending '?per_page=X' if no query parameters exist.
\item Appending '&per_page=X' if other query parameters exist.
\item Overwriting an existing 'per_page' parameter if found.
}
}
\keyword{internal}
