% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utils.R
\name{write_to_renviron}
\alias{write_to_renviron}
\title{Internal function for updating .Renviron variables}
\usage{
write_to_renviron(key, value)
}
\description{
Internal function for updating .Renviron variables
}
\keyword{internal}
