% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_download.R
\name{get_biologer_data}
\alias{get_biologer_data}
\title{Main function used to get data from all logged-in Biologer servers}
\usage{
get_biologer_data(verbose = TRUE)
}
\description{
This function iterates through all servers for which credentials
are stored and downloads field observations, literature observations, and
taxonomic data for each active server.
}
