% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_analysis.R
\name{species_within_polygon}
\alias{species_within_polygon}
\title{Summary of Data within a Polygon}
\usage{
species_within_polygon(
  polygon = NULL,
  country = "all",
  area_buffer = 0,
  region_buffer = 50000,
  auto.download = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{polygon}{An sf or terra spatial object defining the area of interest.}

\item{area_buffer}{Numeric. Buffer distance in meters (default 50).}

\item{region_buffer}{Numeric. Buffer distance in meters (default 50,000).}

\item{auto.download}{Logical. Automatically download new data?}

\item{verbose}{Logical. Print progress messages?}
}
\value{
A summary data frame/table.
}
\description{
Summary of Data within a Polygon
}
